Merges orthogonal_projection.py with polyhedron.py, merges invariant_sets.py with discrete_time_systems.py, adds a notebook with a demo for the MCAIS.
