+++ pympc/dynamics/discrete_time_systems.py
@@ -1,9 +1,10 @@
 # external imports
 import numpy as np
-from scipy.linalg import block_diag
+from scipy.linalg import block_diag, solve_discrete_are
 
 # internal imports
 from pympc.geometry.polyhedron import Polyhedron
+from pympc.optimization.mathematical_programs import LinearProgram
 from pympc.dynamics.discretization_methods import explicit_euler, zero_order_hold
 from pympc.dynamics.utils import check_affine_system
 
@@ -59,6 +60,131 @@ class LinearSystem:
             x.append(self.A.dot(x[-1]) + self.B.dot(v))
 
         return x
+
+    def simulate_closed_loop(self, x0, N, K):
+        """
+        Simulates the system starting from the state x0 for N steps in closed loop with the feedback law u = K x.
+
+        Arguments
+        ----------
+        x0 : numpy.ndarray
+            Initial state.
+        K : numpy.ndarray
+            Feedback gain.
+        N : int
+            Number of simulation steps.
+
+        Returns
+        ----------
+        x : list of numpy.ndarray
+            Sequence of states for t = 0, 1, ..., N.
+        """
+
+        # simulate
+        x = [x0]
+        for t in range(N):
+            x.append((self.A + self.B.dot(K)).dot(x[-1]))
+
+        return x
+
+    def solve_dare(self, Q, R):
+        """
+        Returns the solution of the Discrete Algebraic Riccati Equation (DARE).
+        Consider the linear quadratic control problem V*(x(0)) = min_{x(.), u(.)} 1/2 sum_{t=0}^inf x(t)' Q x(t) + u(t)' R u(t) subject to x(t+1) = A x(t) + B u(t).
+        The optimal solution is u(0) = K x(0) which leads to V*(x(0)) = 1/2 x(0)' P x(0).
+        The pair A, B is assumed to be controllable.
+
+        Arguments
+        ----------
+        Q : numpy.ndarray
+            Quadratic cost for the state (positive semidefinite).
+        R : numpy.ndarray
+            Quadratic cost for the input (positive definite).
+
+        Returns
+        ----------
+        P : numpy.ndarray
+            Hessian of the cost-to-go (positive definite).
+        K : numpy.ndarray
+            Optimal feedback gain matrix.
+        """
+
+        # cost to go
+        P = solve_discrete_are(self.A, self.B, Q, R)
+
+        # feedback
+        K = - np.linalg.inv(self.B.T.dot(P).dot(self.B)+R).dot(self.B.T).dot(P).dot(self.A)
+
+        return P, K
+        
+    def get_mcais(self, X):
+        """
+        See the documentation of mcais().
+        """
+        return mcais(self.A, X)
+
+    def get_mcais_closed_loop(self, K, D):
+        """
+        Returns the maximal constraint-admissible invariant set O_inf for the closed-loop system X(t+1) = (A + B K) x(t).
+        It holds that x(0) in O_inf <=> (x(t), u(t) = K x(t)) in D for all t >= 0.
+
+        Arguments
+        ----------
+        K : numpy.ndarray
+            Stabilizing feedback gain for the linear system.
+        D : instance of Polyhedron
+            Constraint set in the state and input space.
+
+        Returns
+        ----------
+        O_inf : instance of Polyhedron
+            Maximal constraint-admissible (positive) ivariant.
+        t : int
+            Determinedness index.
+        """
+
+        # closed loop dynamics
+        A_cl = self.A + self.B.dot(K)
+
+        # state-space constraint set
+        X_cl = Polyhedron(
+            D.A[:,:self.nx] + D.A[:,self.nx:].dot(K),
+            D.b
+            )
+        O_inf, t = mcais(A_cl, X_cl)
+
+        return O_inf, t
+
+    def get_mcais_closed_loop_orthogonal_domains(self, K, X, U):
+        """
+        Returns the maximal constraint-admissible invariant set O_inf for the closed-loop system X(t+1) = (A + B K) x(t).
+        It holds that x(0) in O_inf <=> x(t) in X and u(t) = K x(t) in U for all t >= 0.
+
+        Arguments
+        ----------
+        K : numpy.ndarray
+            Stabilizing feedback gain for the linear system.
+        X : instance of Polyhedron
+            Constraint set in the state space.
+        U : instance of Polyhedron
+            Constraint set in the input space.
+
+        Returns
+        ----------
+        O_inf : instance of Polyhedron
+            Maximal constraint-admissible (positive) ivariant.
+        t : int
+            Determinedness index.
+        """
+
+        # state- and input-space constraint set
+        D = Polyhedron(
+            block_diag(X.A, U.A),
+            np.vstack((X.b, U.b))
+            )
+        O_inf, t = self.get_mcais_closed_loop(K, D)
+
+        return O_inf, t
 
     def condense(self, N):
         """
@@ -407,6 +533,68 @@ class PieceWiseAffineSystem(object):
             domains.append(Polyhedron(A_i, b_i))
 
         return PieceWiseAffineSystem(affine_systems, domains)
+
+def mcais(A, X, tol=1.e-9):
+    """
+    Returns the maximal constraint-admissible (positive) invariant set O_inf for the system x(t+1) = A x(t) subject to the constraint x in X.
+    O_inf is also known as maximum output admissible set.
+    It holds that x(0) in O_inf <=> x(t) in X for all t >= 0.
+    (Implementation of Algorithm 3.2 from: Gilbert, Tan - Linear Systems with State and Control Constraints, The Theory and Application of Maximal Output Admissible Sets.)
+    Sufficient conditions for this set to be finitely determined (i.e. defined by a finite number of facets) are: A stable, X bounded and containing the origin.
+
+    Arguments
+    ----------
+    A : numpy.ndarray
+        State transition matrix.
+    X : instance of Polyhedron
+        State-space domain of the dynamical system.
+    tol : float
+        Threshold for the checks in the algorithm.
+
+    Returns:
+    ----------
+    O_inf : instance of Polyhedron
+        Maximal constraint-admissible (positive) ivariant.
+    t : int
+        Determinedness index.
+    """
+
+    # ensure convergence of the algorithm
+    eig_max = np.max(np.absolute(np.linalg.eig(A)[0]))
+    if eig_max > 1.:
+        raise ValueError('unstable system, cannot derive maximal constraint-admissible set.')
+    [nc, nx] = X.A.shape
+    if not X.contains(np.zeros((nx, 1))):
+        raise ValueError('the origin is not contained in the constraint set, cannot derive maximal constraint-admissible set.')
+    if not X.bounded:
+        raise ValueError('unbounded constraint set, cannot derive maximal constraint-admissible set.')
+
+    # Gilber and Tan algorithm
+    t = 0
+    convergence = False
+    while not convergence:
+
+        # cost function gradients for all i
+        J = X.A.dot(np.linalg.matrix_power(A,t+1))
+
+        # constraints to each LP
+        F = np.vstack([X.A.dot(np.linalg.matrix_power(A,k)) for k in range(t+1)])
+        g = np.vstack([X.b for k in range(t+1)])
+        O_inf = Polyhedron(F, g)
+
+        # list of all minima
+        J_sol = []
+        lp = LinearProgram(O_inf)
+        for i in range(nc):
+            lp.f = - J[i:i+1,:].T
+            sol = lp.solve()
+            J_sol.append(-sol['min'] - X.b[i])
+        if np.max(J_sol) < tol:
+            convergence = True
+        else:
+            t += 1
+
+    return O_inf, t
 
 def condense_pwa_system(affine_systems, mode_sequence):
     """
